Make the number of days changelog is generated since configurable.
The default value of "CHGLOGDUR" is 10.
If CHGLOGDUR value exceeds 30 it will be forcedly (re)set to 10 in order to avoid performance issues.

To manage the value of "CHGLOGDUR" you can either:
A)add
CHGLOGDUR="<number_of_days>"
after "mka bacon" or "make otapackage" or "whatever-command-you-use-to-start-building-a-rom"
B)input
export CHGLOGDUR="<number_of_days>"